server/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.7
django-braces==1.15.0
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==22.1
django-oauth-toolkit==2.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.4.2
Markdown==3.4.1
oauthlib==3.2.2
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-app-django==5.1.0
social-auth-core==4.4.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.15
wrapt==1.15.0
